fix logic of pattern validator
an Exception is now raised if the value is not integer-valued. happy case return is at the top. the run time of this validator, however, mostly depends on the input data type used (numpy array, vs. list/tuple). the run-time is not crucial here as long as the functionality is ok which seems fine now