New sequencer architecture (#518)
* Initial commit

File with lot of debug stuff and filenames to be probably changed.

* Changed file names and bug fix after initial testing

Created a new file sequencer_widget to start separate widgets in
different files. Renamed and merge some files

* Updated documentation

* Removed redundant code and parent dictionary

* Minor cosmetic changes

* Removed unneeded code

* Added basic tests

* Changed name from SequenceEvaluationException to SequenceEvaluationError

* Removed debug code

* Added interface for EstimatorWidget

* Fixed flake issues

* Moved sequencer_widget file

* Minor fixes

Indentation problem and removed import of SequenceEvaluationException
which is now in a different file

* Removed deprecation message

* Update to new Qt interface

* Fixes for PyQt6/PySide6

* Removed redundant code

* Implemented save method

* Removed debug code

* Fixes for string representation

Thanks to @mcdo0486

* Added load preview and save sequence

Thanks to @mcdo0486

* Fix SequencerTreeView being called with preview. Fix len(filename) on None object.

* Added persistent editor 

Everything not tested.
Other changes: single file selection, load_sequence

* Fixed syntax error

* Added load/append methods

* Changed class name

* Various enhancements

Changed class name
Addressed lint issues
Added checkbox to choose between replace or append.
Added persistent combobox

* Changed class name

Class does not handle file and so name is changed.

* Handle error in files loading

Updated documentation
PyQt6/PySide6 compatibility

* New sequencer (#26)

* Set SequencerTreeModel header arguement to immutable tuple default

* Set default to append_checkbox to checked. Add preview pane when saving files.

* Don't load sequence after hitting cancel

* Set column width of TreeView after setting model

* removed conditional file check on accepted dialog box

* Add docstring, run tests again

* Change valid_inputs argument to immutable tuple. Update tests to reflect additional arguements in SequenceHandler

* Preserve width for SequencerTreeView on init

* Fix for adding item feature.

This has been reported during review and it was a result of a
previous change.

* assert does not need brackets

assert is a keyword, not a function, so it is clearer/more correct to omit the parentheses

* Missing space in docstring

* Removed dunder methods

Removed magic method __len__ and __getitem__ to make code more clear.

* Udpdates tests following changes to SequenceHandler class

* Fix nasty crash when loading sequence without appending

According to Qt documentation, when model data is radically changed,
i.e. when loading new data to replace current one, it is required to
call beginModelReset/endModelReset to invalidate old data.

Previous call to layoutAboutToBeChanged/layoutChanged was not sufficient
and caused a PyQt crash.

* Updated tutorial documentation

Added also fix on sequencer_widget to enable persistent editor upon model reset.

* Update docs/tutorial/graphical.rst

Co-authored-by: Casper Schippers <Casperschippers@hotmail.com>

Co-authored-by: mcdo0486 <mcdo0486@umn.edu>
Co-authored-by: Casper Schippers <Casperschippers@hotmail.com>