Added basic commands to KepcoBOP6312
Added in a minimal set of commands to the KepcoBOP3612 to do some simple measurements. In particular, setting of output mode, voltage and current setpoints and measurement routines.

There are numerous other commands detailed in the manual; however, these are for more sophisticated control and not needed for the moment.