Add _write_bytes() and _read_bytes() for Prologix adapter
As in plain write(), _write_bytes() possibly needs to emit the ++addr
command first. Also, it must always append one CR in order for the
Prologix adapter to actually transfer the data to the target.

In order to read something from a Prologix adapter, it is necessary
to first emit a ++read command, thus the default _read_bytes method
alone is not sufficient.